+++ src/index.ts
@@ -1,2 +1,4 @@
-export { useCountUp } from './hooks'
-export { CountUp } from './components'
+export { useCountUp } from './useCountUp'
+export { CountUp } from './CountUp'
+
+export type { Props, ReturnValue, ReturnProps, Easing, EasingFn } from './types'