+++ src/types.ts
@@ -1,32 +1,9 @@
-import { Options, ElapsedTimeReturnValue } from 'use-elapsed-time'
-
-export type OnComplete = Options['onComplete']
-export type AutoResetKey = Options['autoResetKey']
-export type Reset = ElapsedTimeReturnValue['reset']
+import {
+  ReturnValue as ETReturnValue,
+  Props as ETProps,
+} from 'use-elapsed-time'
 
 export type ReturnValue = number | string | React.ReactNode
-export type Formatter = (value: number) => ReturnValue
-export type ToLocaleStringLocales = string | string[]
-export type ToLocaleStringOptions = {
-  localeMatcher?: string
-  style?: string
-  numberingSystem?: string
-  unit?: string
-  unitDisplay?: string
-  currency?: string
-  currencyDisplay?: string
-  useGrouping?: boolean
-  minimumIntegerDigits?: number
-  minimumFractionDigits?: number
-  maximumFractionDigits?: number
-  minimumSignificantDigits?: number
-  maximumSignificantDigits?: number
-  notation?: string
-}
-export type ToLocaleStringParams = {
-  locale?: ToLocaleStringLocales
-  options?: ToLocaleStringOptions
-}
 
 export type EasingFn = (
   currentTime: number,
@@ -36,14 +13,14 @@ export type EasingFn = (
 ) => number
 export type Easing = 'easeOutCubic' | 'easeInCubic' | 'linear' | EasingFn
 
-export type CountUpReturnProps = {
+export interface ReturnProps {
   /** Current value of the count up animation */
   value: ReturnValue
   /** Method to start over the animation*/
-  reset: Reset
+  reset: ETReturnValue['reset']
 }
 
-export type CountUpProps = {
+export interface Props {
   /** Play and pause counting animation. Default: false */
   isCounting?: boolean
   /** Initial value. Default: 0 */
@@ -58,26 +35,14 @@ export type CountUpProps = {
   decimalSeparator?: string
   /** Thousands separator character. Default: "" */
   thousandsSeparator?: string
-  /** Static text before the value. Default: "" */
-  prefix?: string
-  /** Static text after the value. Default: "" */
-  suffix?: string
-  /** Indicates if toLocaleString should be used. Default: false */
-  shouldUseToLocaleString?: boolean
-  /** Set the toLocaleString locale and options*/
-  toLocaleStringParams?: ToLocaleStringParams
-  /** Static text before the value to be used in case toLocaleString params are not supported. Default: "" */
-  fallbackPrefix?: string
-  /** Static text after the value to be used in case toLocaleString params are not supported. Default: "" */
-  fallbackSuffix?: string
   /** On animation complete event handler */
-  onComplete?: OnComplete
+  onComplete?: ETProps['onComplete']
   /**  Easing function to control how the animation is progressing. Default: easeOutExpo */
   easing?: Easing
   /** Function that formats the output value */
-  formatter?: Formatter
-  /** Render function to render the count up value. Used by the component */
-  children?: (props: CountUpReturnProps) => ReturnValue
-  /** Auto reset animation when the key changes. It works similar to React's key prop */
-  autoResetKey?: AutoResetKey
+  formatter?: (value: number) => ReturnValue
+  /** Update interval in seconds. Determines how often the animated value will change. When set to 0 the value will update on each key frame. Default: 0 */
+  updateInterval?: ETProps['updateInterval']
+  /** On value update event handler. It receives the current value */
+  onUpdate?: (value: ReturnValue) => void
 }