+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 3.0.0 ( ---- , 2021)
+
+**Breaking Changes:**
+
+- IE is not longer supported
+- `autoResetKey` props has been deprecated.
+- `shouldUseToLocaleString` and the whole setup to use the built-in `toLocaleString` out of the box has been deprecated. The same result can be achieved using the `formatter` function.
+- React PropTypes has been removed as well as the `prop-types` peer-dependency. The component and hook will rely on the TypeScript types.
+- `prefix` and `suffix` props has been deprecated. The same result can be achieved just by adding them to in front and behind the value.
+
+**New features:**
+
+- `updateInterval` prop now determines how often the animated value will change. When set to 0 the value will update on each key frame (default behavior).
+- `onUpdate` callback will be fired with the current animated value when it changes.
+
+**Implemented enhancements:**
+
+- bundle size is now even further reduced
+- `esbuild` is now used to bundle the code
+- example folder is added, which can be used for testing the hook and component
+
 ## 2.3.1 (March 25th, 2021)
 
 **Implemented enhancements:**